requirements.txt
fastapi
uvicorn
tensorflow-cpu==2.15.0
numpy<2.0
pillow
python-multipart
